Split test_lol
