+++ tests/testbot.py
@@ -2,7 +2,6 @@
 
 from httpretty import HTTPretty, httprettified
 from plugins.jokes import TofadeEvent
-from mock import patch
 import json
 from common import TestTofbot, TestOrigin, TofbotTestCase
 from common import bot_input, bot_action, bot_kick, set_clock
@@ -73,48 +72,6 @@ class TestCase(TofbotTestCase):
                           "starz de moyenne")
         self.assertOutput('!score alfred', 'alfred: 3.0 starz de moyenne.')
         self.assertOutput('!score michel', "michel n'a pas de starz.")
-
-    def test_lol_kevin(self):
-        self.assertOutput('!kevin', 'pas de Kevin')
-        for msg in ['lol', 'lolerie']:
-            self.bot.send(msg, origin='michel')
-        self.assertOutput('!kevin',
-                          'michel est le Kevin du moment avec 2 lolades')
-        for msg in ['lulz', 'LOL', '10L']:
-            self.bot.send(msg, origin='alfred')
-        self.assertOutput('!kevin',
-                          'alfred est le Kevin du moment avec 3 lolades')
-
-    @patch('plugins.lolrate.datetime_now')
-    def test_lol_rate(self, now_mock):
-
-        self.bot.send('!set lolRateDepth 2')
-        self.assertOutput('!get lolRateDepth', 'lolRateDepth = 2')
-
-        set_clock(now_mock, 12)
-        self.bot.send('lol')
-        self.bot.send('lol')
-        expected = '16 Feb 12h-13h : 2 lolz'
-        self.assertOutput('!lulz', expected)
-        # check that the command itself does not increment
-        self.assertOutput('!lulz', expected)
-
-        set_clock(now_mock, 13)
-        self.bot.send('lol')
-        self.assertOutput('!lulz', ['16 Feb 13h-14h : 1 lolz',
-                                    expected,
-                                    ])
-
-        set_clock(now_mock, 14)
-        self.bot.send('lol')
-        self.assertOutput('!lulz', ['16 Feb 14h-15h : 1 lolz',
-                                    '16 Feb 13h-14h : 1 lolz',
-                                    ])
-
-    def test_lol_kick(self):
-        self.bot.send('lol', origin='michel')
-        l = bot_kick(self.bot)
-        self.assertIn('Au passage, michel est un sacré Kevin', l)
 
     def test_ponce(self):
         del self.bot.plugins["jokes"]
